add: ドラッグのみ可能な状態
