+++ styles/globals.css
@@ -11,6 +11,11 @@ a {
   text-decoration: none;
 }
 
+ul {
+  list-style: none;
+  padding-left: 0;
+}
+
 * {
   box-sizing: border-box;
 }
